feat: update for wallet section footer data binding logic in EditingRedPacketViewController. Add select token scene
